+++ src/compiler/types.c
@@ -515,6 +515,44 @@ bool type_is_int128(Type *type)
 	return kind == TYPE_U128 || kind == TYPE_I128;
 }
 
+Type *type_is_must_init(Type *type)
+{
+	if (!type) return NULL;
+	RETRY:
+	type = type->canonical;
+	switch (type->type_kind)
+	{
+		case TYPE_TYPEDEF:
+			if (type->decl->attr_mustinit) return type;
+			return NULL;
+		case TYPE_OPTIONAL:
+			type = type->optional;
+			goto RETRY;
+		case TYPE_ARRAY:
+		case TYPE_SLICE:
+		case TYPE_FLEXIBLE_ARRAY:
+		case TYPE_INFERRED_ARRAY:
+		case ALL_VECTORS:
+			type = type->array.base;
+			goto RETRY;
+		case TYPE_BITSTRUCT:
+		case TYPE_STRUCT:
+		case TYPE_UNION:
+		{
+			Decl *decl = type->decl;
+			if (decl->attr_mustinit) return type;
+			FOREACH(Decl *, member, decl->strukt.members)
+			{
+				Type *type_inner = type_is_must_init(member->type);
+				if (type_inner) return type_inner;
+			}
+			return NULL;
+		}
+		default:
+			return NULL;
+	}
+}
+
 bool type_is_abi_aggregate(Type *type)
 {
 	return type_is_aggregate(type);