+++ releasenotes.md
@@ -25,6 +25,7 @@
 - Make $sizeof, $alignof and all similar functions return `sz` instead of `usz`.
 - Align literal types with C semantics.
 - Use value promotion instead of signedness promotion to int. So that small unsigned types promote to int, not uint.
+- Add a `@mustinit` attribute to enforce zero-initialization of a type. #3094
 
 ### Stdlib changes
 - `std::collections::RingBuffer` is renamed `RingList`.